+++ src/components/form/components/CheckboxConnectedField.tsx
@@ -1,5 +1,6 @@
 /** @jsx jsx */
 import { jsx } from "@emotion/core"
+import React from "react"
 import { CheckboxFieldBlock } from "./CheckboxFieldBlock"
 import { CheckboxFieldBlockProps } from "./CheckboxFieldBlock"
 import { useConnectedField } from "../hooks/useConnectedField"
@@ -10,10 +11,11 @@ export type CheckboxConnectedFieldProps = {
   label?: React.ReactNode
 } & Omit<CheckboxFieldBlockProps, "id" | "label">
 
-export const CheckboxConnectedField: React.FC<
+export const CheckboxConnectedField = React.forwardRef<
+  HTMLInputElement,
   CheckboxConnectedFieldProps
-> = props => {
+>(function CheckboxConnectedField(props, ref) {
   const [connectedProps] = useConnectedField(props.name)
 
-  return <CheckboxFieldBlock {...connectedProps} {...props} />
-}
+  return <CheckboxFieldBlock ref={ref} {...connectedProps} {...props} />
+})