+++ src/components/form/components/TextAreaConnectedField.tsx
@@ -1,5 +1,6 @@
 /** @jsx jsx */
 import { jsx } from "@emotion/core"
+import React from "react"
 import { TextAreaFieldBlock } from "./TextAreaFieldBlock"
 import { TextAreaFieldBlockProps } from "./TextAreaFieldBlock"
 import { useConnectedField } from "../hooks/useConnectedField"
@@ -10,10 +11,11 @@ export type TextAreaConnectedFieldProps = {
   label?: React.ReactNode
 } & Omit<TextAreaFieldBlockProps, "id" | "label">
 
-export const TextAreaConnectedField: React.FC<
+export const TextAreaConnectedField = React.forwardRef<
+  HTMLTextAreaElement,
   TextAreaConnectedFieldProps
-> = props => {
+>(function TextAreaConnectedField(props, ref) {
   const [connectedProps] = useConnectedField(props.name)
 
-  return <TextAreaFieldBlock {...connectedProps} {...props} />
-}
+  return <TextAreaFieldBlock ref={ref} {...connectedProps} {...props} />
+})