+++ src/typings.d.ts
@@ -8,6 +8,11 @@ declare module "*.svg" {
   export default value
 }
 
+declare module "*.mdx" {
+  let MDXComponent: (props: any) => JSX.Element
+  export default MDXComponent
+}
+
 declare module "gatsby-design-tokens" {
   type MapToString<T> = { [K in keyof T]: string }